+++ app.js
@@ -1,18 +1,45 @@
-const $ = require('jquery');
+$(document).ready(function(){
+  $(function () {
+    $('[data-toggle="tooltip"]').tooltip()
+  });
+  var search = function(userInput, location) {
+    var searchValue = userInput || 'javascript';
+    var locationVal = location || '';
+ 
+    $.ajax({
+      url: 'https://api.meetup.com/find/groups',
+      data: {
+        key: '106175216e3c4ad233449412e3542b',
+        location: locationVal,
+        text: searchValue,
+        upcoming_events: true,
+        radius: 20
+      },
+      dataType: 'jsonp',
+      type: 'GET',
+      success: function(data) {
+        console.log(data);
+        initMap(data.data, locationVal);
+      },
+      error: function(error) {
+        console.log(error);
+      }
+    });
+  }
 
   function initMap(data, location) {
-    const geocoder = new google.maps.Geocoder();
-    let meetups = [
+    var geocoder = new google.maps.Geocoder();
+    var meetups = [
       {name: 'Red', lat: 40.7608, lng: -111.8910},
       {name: 'Green', lat: 40.9608, lng: -111.8910},
       {name: 'Blue', lat: 40.5608, lng: -111.8910},
     ]
     meetups = data;
-    let map = new google.maps.Map(document.getElementById('map'), {
+    map = new google.maps.Map(document.getElementById('map'), {
       center: {lat: 40.7608, lng: -111.8910},
       zoom: 9
     });
-    let address = location;
+    var address = location;
     geocoder.geocode({
       'address': address
     }, function(results, status) {
@@ -22,9 +49,9 @@ const $ = require('jquery');
       console.log('Error reason: ' + status);
     }
     })
-    const infowindow = new google.maps.InfoWindow();
-    for (let i in meetups) {
-      let marker = new google.maps.Marker({
+    var infowindow = new google.maps.InfoWindow();
+    for (var i in meetups) {
+      var marker = new google.maps.Marker({
         position: {lat: meetups[i].lat, lng: meetups[i].lon},
         map: map,
         title: meetups[i].name,
@@ -35,7 +62,7 @@ const $ = require('jquery');
         '<p style="font-family: Play, sans-serif">' + meetups[i].description + '</p>' +
         '<div><a href="' + meetups[i].link + '">' + meetups[i].link + '</div></div>';
       google.maps.event.addListener(marker, 'click', function() {
-        let marker_map = this.getMap();
+        var marker_map = this.getMap();
         infowindow.setContent(this.info)
         infowindow.open(map, this);
       });        
@@ -47,4 +74,13 @@ const $ = require('jquery');
     };
   };
 
-module.exports = initMap;
+  $('form').submit(function(event) {
+    event.preventDefault();
+    $('#content').html('<h1 class="content">List of Meetups</h1>');
+    console.log('Submitted')
+    var input = $('#search-term').val();
+    var location = $('#location').val();
+    search(input, location);
+  });
+  search();
+});
