+++ index.html
@@ -1,39 +1,30 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <meta property="og:locale" content="en_US" />
-    <meta property="og:type" content="website" />
-    <meta property="og:title" content="Meetup Finder"/>
-    <meta property="og:description" content="Find Local Meetup Groups"/>
-    <meta property="og:image" content="https://68.media.tumblr.com/59aef8b4419807e5c2a5728befd0aa41/tumblr_oexph4Om7H1v23hsyo1_500.png">
-    <meta property="og:image:width" content="1187" />
-    <meta property="og:image:height" content="682" />
-    <meta name="twitter:card" content="summary" />
-    <meta name="twitter:title" content="Meetup Finder"/>
-    <meta name="twitter:description" content="Find Local Meetup Groups"/>
-    <!-- <meta name="twitter:site" content="@the_BrianB"/> -->
-    <meta name="twitter:image" content="https://68.media.tumblr.com/59aef8b4419807e5c2a5728befd0aa41/tumblr_oexph4Om7H1v23hsyo1_500.png"/>
-    <!-- <meta name="twitter:creator" content="@the_BrianB"/> -->
-    <title>Meetup Finder Map</title>
+    <title>Simple Map</title>
     <meta name="viewport" content="initial-scale=1.0">
     <meta charset="utf-8">
-    <link href="meetup-finder.min.css" rel="stylesheet">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
+    <link href="https://fonts.googleapis.com/css?family=Francois+One|Play|Exo+2" rel="stylesheet">    
+    <link href='main.css' rel='stylesheet'>
+    <script src='https://ajax.googleapis.com/ajax/libs/jquery/3.1.0/jquery.min.js'></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
     <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyCtiEtpRWLNkFjOgBb1Ykttg49FiqL6ie4"
     async defer></script>
-    <script src='meetup-finder.1.0.0.js'></script>
+    <script src='app.js'></script>    
   </head>
-  <body>
-    <main>
-      <h1 class="header">Find a Local Meetup Group</h1>
+  <body>   
+    <main> 
+      <h1 class="header">Find a Local Meetup Group</h1>      
       <form id="search">
         <input id="search-term" type="text" placeholder="Meetup Topic" data-toggle="tooltip" data-placement="left" title="Eg: Tennis, Fishing, Golf, Pokémon, etc.">
         <input id="location" type="text" placeholder="City" data-toggle="tooltip" data-placement="right" title="Eg: Los Angeles, New York, Salt Lake City, Seattle, etc.">
         <button type="submit" form="search" class="search" value="Search">Search</button>
       </form>
     </main>
-    <div id="map"></div>
+    <div id="map"></div> 
     <div id="content">
       <h1 class="content">List of Meetups</h1>
-    </div>
+    </div>  
   </body>
 </html>
