--- meetup-finder.1.0.0.js
@@ -1,20 +1,0 @@
-var $ = require('jquery');
-var search = require('./search');
-
-$(document).ready(function(){
-  $(function () {
-    $('[data-toggle="tooltip"]').tooltip()
-  });
-
-
-  $('form').submit(function(event) {
-    event.preventDefault();
-    $('#content').html('<h1 class="content">List of Meetups</h1>');
-    console.log('Submitted')
-    var input = $('#search-term').val();
-    var location = $('#location').val();
-    search(input, location);
-  });
-  search();
-});
-
