--- package.json
@@ -1,50 +1,0 @@
-{
-  "name": "meetup-finder",
-  "version": "1.0.0",
-  "description": "Meetup Finder\r ======",
-  "main": "meetup-finder.1.0.0.js",
-  "scripts": {
-    "mkdir": "mkdir build",
-    "build": "npm run clean && npm run mkdir && npm run build:html && npm run build:css && npm run build:js",
-    "watch": "npm run watch:html & npm run watch:css & npm run watch:js",
-    "clean": "rm -rf build",
-    "build:html": "npm run clean:html && cp index.html build/",
-    "clean:html": "rm -f build/index.html",
-    "watch:html": "npm run build:html && chokidar index.html -c 'npm run build:html'",
-    "build:js": "npm run clean:js && webpack",
-    "clean:js": "rm -f build/meetup-finder.1.0.0.js build/meetup-finder.1.0.0.js.map",
-    "watch:js": "npm run build:js && webpack --watch",
-    "build:css": "npm run build:css:unminified && npm run build:css:minified",
-    "watch:css": "npm run watch:css:unminified & npm run watch:css:minified",
-    "clean:css": "npm run clean:css:unminified && npm run clean:css:minified",
-    "build:css:unminified": "npm run clean:css:unminified && lessc --source-map styles/main.less build/meetup-finder.css",
-    "clean:css:unminified": "rm -f build/meetup-finder.css build/meetup-finder.css.map",
-    "watch:css:unminified": "npm run build:css && chokidar 'styles/**/*.less' -c 'npm run build:css'",
-    "build:css:minified": "npm run clean:css:minified && lessc --clean-css --source-map styles/main.less build/meetup-finder.min.css",
-    "clean:css:minified": "rm -f build/meetup-finder.min.css build/meetup-finder.min.css.map",
-    "watch:css:minified": "npm run build:css && chokidar 'styles/**/*.less' -c 'npm run build:css'"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/briancbarrow/meetup-finder.git"
-  },
-  "author": "Brian Barrow",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/briancbarrow/meetup-finder/issues"
-  },
-  "homepage": "https://github.com/briancbarrow/meetup-finder#readme",
-  "dependencies": {
-    "bootstrap": "^3.3.7",
-    "jquery": "^3.1.1"
-  },
-  "devDependencies": {
-    "babel-core": "^6.17.0",
-    "babel-loader": "^6.2.5",
-    "babel-preset-es2015": "^6.16.0",
-    "chokidar-cli": "^1.2.0",
-    "less": "^2.7.1",
-    "less-plugin-clean-css": "^1.5.1",
-    "webpack": "^1.13.2"
-  }
-}
