--- styles/main.less
@@ -1,89 +1,0 @@
-@main-background: #FAFFF9;
-@main-text: #006032;
-@list-font: 'Play', sans-serif;
-@main-font: 'Francois One', sans-serif;
-
-html, body {
-  height: 85%;
-  width: 80%;
-  margin: 0 auto;
-  padding: 0;
-  text-align: center;
-  background-color: @main-background;
-}
-#map {
-  height: 100%;
-  border: solid 1px #183059;
-}
-
-main {
-  margin-bottom: 15px;
-  color: @main-text;
-  font-family: @main-font;
-}
-
-.header {
-  display: inline-block;
-  margin-top: 5px;
-}
-
-.hint {
-  display: inline;
-  font-size: 15px;
-}
-
-input {
-  background-color: #F6F4F3;
-  padding-left: 5px;
-  border-top: none;
-  border-left: none;
-  border-right: none;
-  border-bottom: 1px solid #32936F;
-  background-color: @main-background;
-  font-size: 20px;
-  margin: 5px;
-  &:focus {
-    outline: none;
-  }
-  &::-webkit-input-placeholder { /* Chrome/Opera/Safari */
-    color: #AFBFAC;
-  }
-  &::-moz-placeholder { /* Chrome/Opera/Safari */
-    color: #AFBFAC;
-  }
-  &:-ms-input-placeholder { /* Chrome/Opera/Safari */
-    color: #AFBFAC;
-  }
-  &:-moz-placeholder { /* Chrome/Opera/Safari */
-    color: #AFBFAC;
-  }
-}
-
-button {
-  display: block;
-  margin: 5px auto 15px auto;  
-  font-size: 20px;
-  border-radius: 9px;
-  background-color: #32936F;
-  height: 35px;
-  border-color: #596869;
-  color: #303328;
-  &:focus {
-    outline: none;
-  }
-}
-
-.meetup {
-  border: solid 1px @main-text;
-  text-align: left;
-  margin: 10px auto;
-  padding: 0 10px 10px 10px;
-  font-family: @list-font;
-}
-
-.content {
-  font-family: @main-font;
-  color: @main-text;
-  text-align: left;
-  margin-top: 35px;
-}
